requirements.txt
asgiref==3.7.2
boto3==1.28.0
botocore==1.31.0
Django==4.2.3
django-ses==3.5.0
django-utils-six==2.0
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
numpy==1.25.0
Pillow==10.0.0
psycopg2-binary==2.9.6
pycountry==22.3.5
pydicom==2.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
